Move runBatchDryRun from Table to App
